+++ src/Entity/CommentEntity.php
@@ -2,15 +2,16 @@
 
 namespace App\Entity;
 
-class CommentsEntity
+class CommentEntity
 {
     public ?int $id;
     public ?int $uid;
     public ?int $bid;
-    public ?string $name = '';
-    public ?string $mail = '';
-    public ?string $surname = '';
-    public ?string $txt = '';
+    public ?string $name;
+    public ?string $mail;
+    public ?string $auth;
+    public ?string $surname;
+    public ?string $txt;
     public ?int $pub;
     public ?string $date;
 
@@ -34,6 +35,10 @@ class CommentsEntity
     {
         return $this->mail;
     }
+    public function getAuth(): string
+    {
+        return $this->auth;
+    }
     public function getSurname(): string
     {
         return $this->surname;
@@ -46,7 +51,7 @@ class CommentsEntity
     {
         return $this->pub;
     }
-    public function getDate(): int
+    public function getDate(): string
     {
         return $this->date;
     }