+++ src/Repository/CategoryRepository.php
@@ -4,20 +4,18 @@ declare(strict_types=1);
 
 namespace App\Repository;
 
-use App\Model\MainPdo;
 use App\Entity\CategoryEntity;
-use App\Model\Connect;
-use PDO;
+use App\Service\DbService;
 
-class CategoryRepository extends MainPdo
+class CategoryRepository
 {
     protected static string $table = 'cats';
     private static $instance;
-    private Connect $connect;
+    private DbService $dbService;
 
     private function __construct()
     {
-        $this->connect = Connect::getInstance();
+        $this->dbService = DbService::getInstance();
     }
 
     public static function getInstance(): self
@@ -31,21 +29,13 @@ class CategoryRepository extends MainPdo
     public function allCategories(): array
     {
         $sql = 'select id,category from ' . self::$table . '';
-        $pdo = $this->connect->pdo;
-        $stmt = $pdo->prepare($sql);
-        $stmt->setFetchMode(PDO::FETCH_CLASS, CategoryEntity::class, null);
-        $stmt->execute();
-        return $stmt->fetchAll();
+        return $this->dbService->fetchAllCategories($sql, []);
     }
 
     public function findCategoryFromId(int $id): CategoryEntity
     {
         $sql = 'select category from ' . self::$table . ' where id=?';
-        $pdo = $this->connect->pdo;
-        $stmt = $pdo->prepare($sql);
-        $stmt->setFetchMode(PDO::FETCH_CLASS, CategoryEntity::class, null);
-        $stmt->execute([$id]);
-        return $stmt->fetch();
+        return $this->dbService->fetchCategory($sql, [$id]);
     }
 
 }