+++ src/Repository/userRepository.php
@@ -4,21 +4,19 @@ declare(strict_types=1);
 
 namespace App\Repository;
 
-use App\Model\MainPdo;
 use App\Entity\UserEntity;
-use App\Model\Connect;
-use PDO;
+use App\Service\DbService;
 
-class UserRepository extends MainPdo
+class UserRepository
 {
     protected static string $table = 'users';
     protected static string $socials = 'socials';
     private static $instance;
-    private Connect $connect;
+    private DbService $dbService;
 
     private function __construct()
     {
-        $this->connect = Connect::getInstance();
+        $this->dbService = DbService::getInstance();
     }
 
     public static function getInstance(): self
@@ -36,57 +34,31 @@ class UserRepository extends MainPdo
         left join profile
         on users.id=uid
         where users.id=?';
-        $pdo = $this->connect->pdo;
-        $stmt = $pdo->prepare($sql);
-        $stmt->setFetchMode(PDO::FETCH_CLASS, UserEntity::class, null);
-        $stmt->execute([$id]);
-        return $stmt->fetch();
+        return $this->dbService->fetchUser($sql, [$id]);
     }
 
     public function findUserFromId(int $id): UserEntity
     {
         $sql = 'select name from ' . self::$table . ' where id=?';
-        $pdo = $this->connect->pdo;
-        $stmt = $pdo->prepare($sql);
-        $stmt->setFetchMode(PDO::FETCH_CLASS, UserEntity::class, null);
-        $stmt->execute([$id]);
-        return $stmt->fetch();
+        return $this->dbService->fetchUser($sql, [$id]);
     }
 
     public function findUserFromName(string $name): UserEntity|bool
     {
         $sql = 'select id,name,pswd from ' . self::$table . ' where name=?';
-        $pdo = $this->connect->pdo;
-        $stmt = $pdo->prepare($sql);
-        $stmt->setFetchMode(PDO::FETCH_CLASS, UserEntity::class, null);
-        $stmt->execute([$name]);
-        $array = $stmt->fetch();
-        //vd($stmt);
-        return $array;
+        return $this->dbService->fetchUser($sql, [$name]);
     }
 
-    public function registerUser(array $values): string
+    public function registerUser(array $blind): string
     {
         $sql = 'insert into ' . self::$table . ' values (null, :name, 1, :mail, :pswd, now())';
-        $pdo = $this->connect->pdo;
-        $stmt = $pdo->prepare($sql);
-        $stmt->setFetchMode(PDO::FETCH_CLASS, UserEntity::class, null);
-        //$stmt->bindParam(':name', $name, PDO::PARAM_STR);
-        //$stmt->bindParam(':mail', $mail, PDO::PARAM_STR);
-        //$stmt->bindParam(':pswd', $pswd, PDO::PARAM_STR);
-        $stmt->execute($values);
-        $id = $pdo->lastInsertId();
-        return $id;
+        return $this->dbService->fetchAddUser($sql, $blind);
     }
 
     public function userLinks(int $id): array
     {
         $sql = 'select url from socials where uid=?';
-        $pdo = $this->connect->pdo;
-        $stmt = $pdo->prepare($sql);
-        $stmt->setFetchMode(PDO::FETCH_CLASS, UserEntity::class, null);
-        $stmt->execute([$id]);
-        return $stmt->fetchAll();
+        return $this->dbService->fetchAllSocials($sql, [$id]);
     }
 
 }